requirements.txt
zipfile
gdown
streamlit
Pillow
numpy
tensorflow